refactor(*): Rename Future.Cancel to Drop
The renaming was done due to the fact that it is mandatory not only for the semantic cancellation of asynchronous computation, but also for cleaning up resources and notifying the waiting source in case of an exception in the associated Future. For example, Merge should stop f2 if an exception has occurred in f1.
It looks strange if you remember that the exception is a form of the result. But in this form, the cancellation call would be optional. But in fact it is technically necessary.
Therefore, a semantic change in the design of Future.Cancel.